+++ library/klip-core/src/commonMain/kotlin/int/KlipManager.kt
@@ -5,7 +5,12 @@ import dev.petuska.klip.core.ext.readText
 import dev.petuska.klip.core.ext.writeText
 import kotlin.native.concurrent.ThreadLocal
 
-public typealias Klips = MutableMap<String, String>
+public typealias Klips = MutableMap<String, Klip>
+
+public data class Klip(
+  val value: String,
+  val attributes: Map<String, String>,
+)
 
 /**
  * A helper class to assist in reading and writing persisted klips
@@ -13,13 +18,15 @@ public typealias Klips = MutableMap<String, String>
 @ThreadLocal
 public object KlipManager {
   private const val SEPARATOR = ":::::>>"
+  private const val SEPARATOR_ATTR = "="
+  private const val SEPARATOR_ATTRS = ";"
   private const val SEPARATOR_KEY = "$SEPARATOR\n"
   private const val SEPARATOR_KLIPS = "\n$SEPARATOR"
   private const val SEPARATOR_SOF = "$SEPARATOR KLIPS $SEPARATOR"
   private const val SEPARATOR_EOF = "\n$SEPARATOR$SEPARATOR$SEPARATOR\n"
   private val klipMatrix = mutableMapOf<String, Klips>()
 
-  private fun loadKlips(path: String) = klipMatrix[path] ?: run {
+  private fun loadKlips(path: String): Klips = klipMatrix[path] ?: run {
     val klips = read(path)
       ?.replace("\r\n", "\n")
       ?.removePrefix(SEPARATOR_SOF)
@@ -32,7 +39,12 @@ public object KlipManager {
           "Corrupted klip at $path:${split.getOrNull(0)?.substringBefore(SEPARATOR)}"
         }
         val (k, v) = split
-        k to v
+        val (attr, key) = k.split(SEPARATOR).let { if (it.size > 1) it[0] to it[1] else "" to it[0] }
+        val attributes = attr.takeIf(String::isNotBlank)
+          ?.split(SEPARATOR_ATTRS)
+          ?.associate { it.split(SEPARATOR_ATTR).let { (k, v) -> k to v } }
+          ?: mapOf()
+        key to Klip(v, attributes)
       }?.toMutableMap() ?: mutableMapOf()
     klips.also {
       klipMatrix[path] = it
@@ -46,7 +58,9 @@ public object KlipManager {
         separator = "",
         prefix = SEPARATOR_SOF,
         postfix = SEPARATOR_EOF
-      ) { (k, v) -> stringifyKlip(k) { v } }
+      ) { (k, v) ->
+        stringifyKlip(k, v.attributes) { v.value }
+      }
     }
   }
 
@@ -69,10 +83,11 @@ public object KlipManager {
   /**
    * Converts a [key] and a klip provided by [source] to a writable form that's safe to append to klip file.
    * @param key a key to identify the klip within the file
+   * @param attributes klip attributes
    * @return klip-writable form of the value provided by [source] that's safe to append to klip file
    */
-  public fun stringifyKlip(key: String, source: () -> String): String {
-    return "$SEPARATOR_KLIPS$key$SEPARATOR_KEY${source()}"
+  public fun stringifyKlip(key: String, attributes: Map<String, String>, source: () -> String): String {
+    return "${SEPARATOR_KLIPS}${attributes.entries.joinToString { (k, v) -> "$k=$v" }}$SEPARATOR$key$SEPARATOR_KEY${source()}"
   }
 
   /**
@@ -80,12 +95,13 @@ public object KlipManager {
    *
    * @param path file path to retrieve and save klip for the given context
    * @param key key to write the klip under in the file at [path]
+   * @param attributes klip attributes
    * @param source a provider for a string representation to persist
    * @return value provided by [source]
    */
-  public fun writeKlip(path: String, key: String, source: () -> String): String {
+  public fun writeKlip(path: String, key: String, attributes: Map<String, String>, source: () -> String): Klip {
     val klips = loadKlips(path)
-    return source().also { saveKlips(path, (klips + mutableMapOf(key to it)).toMutableMap()) }
+    return Klip(source(), attributes).also { saveKlips(path, (klips + mutableMapOf(key to it)).toMutableMap()) }
   }
 
   /**
@@ -94,15 +110,16 @@ public object KlipManager {
    *
    * @param path file path to retrieve and save klip for the given context
    * @param key key to read or write the klip under in the file at [path]
+   * @param attributes klip attributes
    * @param source a provider for a string representation to persist
    * @return read klip if it exists or persisted value that was provided by [source] otherwise
    */
-  public fun readKlip(path: String, key: String, source: () -> String): String {
+  public fun readKlip(path: String, key: String, attributes: Map<String, String>, source: () -> String): Klip {
     val klips = loadKlips(path)
     val klip = klips[key]
     return when {
       klip != null -> klip
-      else -> writeKlip(path, key, source)
+      else -> writeKlip(path, key, attributes, source)
     }
   }
 
@@ -110,14 +127,15 @@ public object KlipManager {
    * Writes or returns klip depending on the working mode.
    * @param context [KlipContext] containing metadata about the klip
    * @param source a provider for a string representation to persist
+   * @param attributes klip attributes
    * @return read klip if it exists and [KlipContext.update]` == false`
    * or persisted value that was provided by [source] otherwise
    */
-  public fun klip(context: KlipContext, source: () -> String): String = with(context) {
+  public fun klip(context: KlipContext, attributes: Map<String, String>, source: () -> String): Klip = with(context) {
     return if (update) {
-      writeKlip(path, key, source)
+      writeKlip(path, key, attributes, source)
     } else {
-      readKlip(path, key, source)
+      readKlip(path, key, attributes, source)
     }
   }
 }