+++ library/klip-core/src/commonMain/kotlin/int/KlipType.kt
@@ -1,0 +1,5 @@
+package dev.petuska.klip.core.int
+
+public enum class KlipType {
+  TEXT, JSON
+}