+++ plugin/klip-kotlin-plugin/src/main/kotlin/KlipComponentRegistrar.kt
@@ -6,7 +6,7 @@ import dev.petuska.klip.plugin.util.KlipSettings
 import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.cli.common.CLIConfigurationKeys
 import org.jetbrains.kotlin.cli.common.messages.MessageCollector
-import org.jetbrains.kotlin.cli.js.messageCollectorLogger
+import org.jetbrains.kotlin.cli.common.toLogger
 import org.jetbrains.kotlin.com.intellij.mock.MockProject
 import org.jetbrains.kotlin.compiler.plugin.ComponentRegistrar
 import org.jetbrains.kotlin.config.CompilerConfiguration
@@ -37,8 +37,10 @@ class KlipComponentRegistrar : ComponentRegistrar {
                     scopeAnnotations =
                         configuration[KlipOption.ScopeAnnotation.key]?.map { FqName(it) }
                             ?: listOf(),
+                    scopeFunctions = configuration[KlipOption.ScopeFunction.key]?.map { FqName(it) }
+                            ?: listOf(),
                 ),
-            logger = messageCollectorLogger(messageCollector),
+            logger = messageCollector.toLogger(),
         ))
   }
 }