+++ sandbox/src/jvmTest/kotlin/kotest/JvmWordSpec.kt
@@ -1,0 +1,19 @@
+package sandbox.test.kotest
+
+import dev.petuska.klip.api.assertKlip
+import dev.petuska.klip.api.assertMatchesKlip
+import io.kotest.core.spec.style.WordSpec
+
+class JvmWordSpec : WordSpec({
+  "Hello" When {
+    "context" should {
+      "test one" {
+        assertMatchesKlip("kotest zero")
+        assertMatchesKlip("kotest one")
+      }
+      "test two" {
+        "kotest 2".assertKlip()
+      }
+    }
+  }
+})